lab 17 p2

+++ 17_servlets_jsp/servlets_task2/src/main/java/com/nixsolutions/dao/BookDao.java
@@ -9,7 +9,7 @@ public interface BookDao {
 	public List<Book> getBooksByAuthor(String name);
 	public List<Book> getBooksByCategory(String name);
 	public List<Book> getBooksByName(String name);
-	public Book getBookById(int bookId);
+	public Book getBookById(Long bookId);
 	public Book createBook(Book book);
 	public void updateBook(Book book);
 	public void deleteBook(Book book);
