+++ 17_servlets_jsp/servlets_task2/src/main/java/com/nixsolutions/entity/Category.java
@@ -1,7 +1,7 @@
 package com.nixsolutions.entity;
 
 public class Category {
-	private Integer categoryId;
+	private Long categoryId;
 	private String name;
 
 	public Category(String name) {
@@ -9,11 +9,11 @@ public class Category {
 		this.name = name;
 	}
 
-	public Integer getCategoryId() {
+	public Long getCategoryId() {
 		return categoryId;
 	}
 
-	public void setCategoryId(int categoryId) {
+	public void setCategoryId(Long categoryId) {
 		this.categoryId = categoryId;
 	}
